Extract ``changes`` into a new extension
